requirements.txt
torch==2.0.1
torchaudio==2.0.2
pyannote.audio==3.1.1
huggingface_hub==0.17.3
openai-whisper
edge-tts
pydub
streamlit
requests
numpy<2.0.0
